Bugs eat rat
BUGS EAT RATS
The idea 💡  of ​​the program is that you put the file 📁  that you suspect and it is checked for viruses 🐀  , the IP and TCP are checked and it is checked to see if it is modified or something like that.
Do not rely on the program 100% because it only scans files. It is likely to be better than Windows scanning because it scans the IP and TPC of the program.
The program is free and not protected and it is possible to tamper with it and sell it or add things like viruses. Please take it from its original source which is here on this server.
The program is something new. If there are any errors, please let me know. It is likely that the program is not being reported correctly.